[inductor] Add single pass "var" reduction_type
This is a bit inefficient because it computes the mean but thows it
away since ir.Reduction nodes only have 1 output. However, the mean
can at least be scheduled into the same loop as the variance now since
there is no data dependency. Thus we can take fewer passes over the
data.

ghstack-source-id: b8b0df9f7dfd89599eb8cbff3d5bd563994a718a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102486